FlutterViewDelegate to own class
