+++ shell/platform/android/io/flutter/embedding/android/FlutterView.java
@@ -1450,34 +1450,6 @@ public class FlutterView extends FrameLayout
         .send();
   }
 
-  /** A delegate class that performs the task of retrieving the bounding rect values. */
-  public static class FlutterViewDelegate {
-
-    /**
-     * Return the WindowInsets object for the provided Context, or null if there is no associated
-     * activity.
-     */
-    @RequiresApi(api = API_LEVELS.API_23)
-    @VisibleForTesting
-    public WindowInsets getWindowInsets(Context context) {
-      Activity activity = ViewUtils.getActivity(context);
-      if (activity == null) {
-        return null;
-      }
-      return activity.getWindow().getDecorView().getRootWindowInsets();
-    }
-
-    @RequiresApi(api = API_LEVELS.API_35)
-    @VisibleForTesting
-    public List<Rect> getCaptionBarInsets(Context context) {
-      WindowInsets insets = getWindowInsets(context);
-      if (insets == null) {
-        return Collections.emptyList();
-      }
-      return insets.getBoundingRects(WindowInsets.Type.captionBar());
-    }
-  }
-
   private FlutterViewDelegate delegate = new FlutterViewDelegate();
 
   /** Set the FlutterViewDelegate, such as to a mock for testing. */
